+++ tests/testthat/test_q.R
@@ -5,6 +5,8 @@ context("capture")
 test_that("test_q: Works As Expected", {
   d <- table_source("d", c("AUC", "R2"))
   v <- "NEWVAR"
+  AUC <- "NEWVAR"
+  R2 <- "NEWVAR"
   optree <- extend_nse(d, v %:=% R2 + 1)
   str <- format(optree)
   testthat::expect_equal(0, length(grep("NEWVAR", str, fixed = TRUE)))